+++ src/main/cn/edu/uestc/ostec/workload/controller/core/ApplicationController.java
@@ -33,7 +33,7 @@ public class ApplicationController extends ResultController
 		return servletContext.getContextPath();
 	}
 
-	public Long getUserId() {
+	public Integer getUserId() {
 		return getUser().getUserId();
 	}
